+++ tests/test_mesh.py
@@ -323,6 +323,18 @@ def test_container():
         print(container.as_meshio(combined=combined))
 
 
+def test_read(filename="tests/mesh.bdf"):
+
+    mesh = fe.mesh.read(filename=filename, dim=2)[0]
+    assert mesh.dim == 2
+
+    mesh = fe.mesh.read(filename=filename, dim=None)[0]
+    assert mesh.dim == 3
+
+    mesh = fe.mesh.read(filename=filename, cellblock=0)[0]
+    assert mesh.dim == 3
+
+
 if __name__ == "__main__":
     test_meshes()
     test_mirror()
@@ -332,3 +344,4 @@ if __name__ == "__main__":
     test_grid()
     test_grid_1d()
     test_container()
+    test_read(filename="mesh.bdf")